feat(#1711): fix flaky test

+++ eo-maven-plugin/src/main/java/org/eolang/maven/SafeMojo.java
@@ -326,13 +326,13 @@ abstract class SafeMojo extends AbstractMojo {
                     this.exec();
                     return new Object();
                 }
-            ).get(this.timeout, TimeUnit.SECONDS);
+            ).get(this.timeout.longValue(), TimeUnit.SECONDS);
         } catch (final InterruptedException ex) {
             Thread.currentThread().interrupt();
             throw new IllegalStateException(
                 Logger.format(
                     "Timeout %[ms]s thread was interrupted",
-                    TimeUnit.SECONDS.toMillis(this.timeout)
+                    TimeUnit.SECONDS.toMillis(this.timeout.longValue())
                 ),
                 ex
             );
